printf

+++ ft_printf.h
@@ -21,7 +21,7 @@
 int	ft_printf(char *format, ...);
 int	ft_putchar(char c, int *count);
 int	ft_putstr(char *str, int *count);
-int	ft_putnbr_base(unsigned long long nbr, char *base, int base_len,
+int	ft_putnbr_base(unsigned long int nbr, char *base, int base_len,
 		int *count);
 int	ft_putpointer(void *ptr, int *count);
 int	ft_putnbr_signed(int nbr, int *count);
